Aplicacion de consola de ejemplo

+++ HuellaCarbono.LogicaNegocios/modelos/Actividad.cs
@@ -6,7 +6,7 @@ namespace HuellaCarbono.LogicaNegocios.modelos
     {
         public int Id { get; set; }
         public String Descripcion { get; set; }
-        public decimal FactorEmision { get; set; }
+        public Double FactorEmision { get; set; }
         public Unidad UnidadDeActividad { get; set; }
     }
 }
